MA-9041 [CTS]Modify the package imxCts to adapt 7.1_cts
The latest cts should run with -m module_name, which
requires a relevant package_name.config file storing
in testcases folder.
To create this config file automatically:
1. Add AndroidTest.xml at root directory of the package project.
2. Add LOCAL_COMPATIBILITY_SUITE := cts to Android.mk

For Android 7.0 and higher, we should not use Uri.fromFile()
API to return a content://URI, which will result a
FileUriExposedException due to Android 7.0's strictMode.
As an alternative, chose FileProvider to get the file URI.

To run this module, some preconditions should be done:
1. cd maddev_ng7.1/packages/apps/fsl_imx_demo/imxCts
2. mma
3. cp maddev_ng7.1/out/host/linux-x86/cts/android-cts
   /testcases/CtsImxTestCases.* android-cts/testcases/
4. Set androidboot.selinux=permissive to give the test
   permissions to access fb node.
5. Make sure sdcard has cts test media files.
6. Make sure camera works normal.
7. Make sure network is connected.
8. Disable WebView Broswer in settings->apps, and
   install Chrome apk instead.
And then run cts -m CtsImxTestCases.

Change-Id: I171699e85931e43045b68690c9a0556976d52bf0
Signed-off-by: Jindong <jindong.yue@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/1341
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>